+++ frickform/ControlExtensions.cs
@@ -1,4 +1,4 @@
-﻿namespace frickative
+﻿namespace frickform
 {
     public static class ControlExtensions
     {
