+++ frickform/VerticalLabel.cs
@@ -1,6 +1,6 @@
 ﻿using System.Drawing.Drawing2D;
 
-namespace frickative
+namespace frickform
 {
     public class VerticalLabel : Label
     {
